# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reservations - API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/cinemas/{cinemaId}/rooms/{roomId}/seances": {
            "get": {
                "tags": [
                    "Seances"
                ],
                "summary": "Gets the list of seances by room",
                "description": "Returns the list of seances by room with optional pagination",
                "operationId": "getSeancesListByRoom",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of cinemas per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "cinemaId",
                        "in": "path",
                        "description": "ID of cinema to return rooms",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "roomId",
                        "in": "path",
                        "description": "ID of room to return seances",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No seance(s) found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Seances"
                ],
                "summary": "Store a new seance by room",
                "description": "Stores a new seance by room and returns the seance data",
                "operationId": "storeSeanceByRoom",
                "parameters": [
                    {
                        "name": "cinemaId",
                        "in": "path",
                        "description": "ID of cinema to store room",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "roomId",
                        "in": "path",
                        "description": "ID of room to store seance",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Seance data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "movie_id",
                                    "start_date"
                                ],
                                "properties": {
                                    "movie_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "3434-4343-4343-4343"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-09-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "3434-4343-4343-4343"
                                        },
                                        "movie_id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "3434-4343-4343-4343"
                                        },
                                        "start_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        },
                                        "room_id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "3434-4343-4343-4343"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input data"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    }
}